requirements.txt
pandas>=1.3.0
openpyxl>=3.0.7
PyYAML>=5.4.1
pywin32>=300 